Pada filter semua kelas jadwal tetap berupa jadwal interaktif

+++ src/app/jadwal/page.tsx
@@ -160,7 +160,7 @@ export default function JadwalPage() {
   };
   
   const confirmDelete = () => {
-    if (jadwalToDelete) {
+    if (jadwalToDelete && firestore) {
         const jadwalDocRef = doc(firestore, 'jadwal', jadwalToDelete.id);
         deleteDocumentNonBlocking(jadwalDocRef);
         setJadwalToDelete(null);
